+++ app/Stubs/Migrations/HasFulfilmentStats.php
@@ -20,13 +20,13 @@ trait HasFulfilmentStats
         $table->unsignedInteger('number_stored_items')->default(0);
 
         foreach (StoredItemTypeEnum::cases() as $type) {
-            $table->unsignedBigInteger("number_stored_items_type_{$type->snake()}")->default(0);
+            $table->unsignedInteger("number_stored_items_type_{$type->snake()}")->default(0);
         }
         foreach (StoredItemStateEnum::cases() as $state) {
-            $table->unsignedBigInteger("number_stored_items_state_{$state->snake()}")->default(0);
+            $table->unsignedInteger("number_stored_items_state_{$state->snake()}")->default(0);
         }
         foreach (StoredItemStatusEnum::cases() as $status) {
-            $table->unsignedBigInteger("number_stored_items_status_{$status->snake()}")->default(0);
+            $table->unsignedInteger("number_stored_items_status_{$status->snake()}")->default(0);
         }
 
         return $table;
@@ -38,10 +38,10 @@ trait HasFulfilmentStats
         $table->unsignedInteger('number_customers_with_assets')->default(0);
 
         foreach (StoredItemStateEnum::cases() as $state) {
-            $table->unsignedBigInteger("number_customers_with_stored_items_state_{$state->snake()}")->default(0);
+            $table->unsignedInteger("number_customers_with_stored_items_state_{$state->snake()}")->default(0);
         }
         foreach (StoredItemStatusEnum::cases() as $status) {
-            $table->unsignedBigInteger("number_customers_with_stored_items_status_{$status->snake()}")->default(0);
+            $table->unsignedInteger("number_customers_with_stored_items_status_{$status->snake()}")->default(0);
         }
 
         return $table;
