+++ database/migrations/2022_09_01_104900_create_products_table.php
@@ -34,7 +34,7 @@ return new class () extends Migration {
             $table->unsignedDecimal('price', 18)->comment('unit price');
             $table->unsignedDecimal('rrp', 12, 3)->nullable()->comment('RRP per outer');
             $table->unsignedInteger('available')->default(0)->nullable();
-            $table->unsignedBigInteger('image_id')->nullable();
+            $table->unsignedInteger('image_id')->nullable();
             $table->jsonb('settings');
             $table->jsonb('data');
             $table->timestampsTz();
