+++ database/migrations/2023_04_10_080134_media_stock_family.php
@@ -16,7 +16,7 @@ return new class () extends Migration {
             $table->unsignedInteger('stock_family_id')->index();
             $table->string('type')->index();
             $table->foreign('stock_family_id')->references('id')->on('stock_families');
-            $table->unsignedBigInteger('media_id')->index();
+            $table->unsignedInteger('media_id')->index();
             $table->unique(['stock_family_id', 'media_id']);
             $table->string('owner_type')->index();
             $table->unsignedInteger('owner_id');
